Template: only template, updated all image Paths
Changed the folder structure so that only the template project is here.
updated template project. the default template npm run build wasn't working as it was using the old way of using images paths, i put everything in public folder and updated the paths.
There might be repercussions as i used AI cursor to update the paths, so i will have to watch out because some .js files that were storing paths, are now being overritten in the component with the path directly instead of using the stored variable of the path